Move `isAbortError` to error.ts
