--- web/packages/shared/utils/abortError.test.ts
@@ -1,45 +1,0 @@
-/**
- * Teleport
- * Copyright (C) 2023  Gravitational, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { isAbortError } from 'shared/utils/abortError';
-
-describe.each([
-  ['DOMException', newDOMAbortError],
-  ['ApiError', newApiAbortError],
-  ['gRPC Error', newGrpcAbortError],
-])('for error type %s', (_, ErrorType) => {
-  it('is abort error', async () => {
-    expect(isAbortError(ErrorType())).toBe(true);
-  });
-});
-
-function newDOMAbortError() {
-  return new DOMException('Aborted', 'AbortError');
-}
-
-// mimics ApiError
-function newApiAbortError() {
-  return new Error('The user aborted a request', {
-    cause: newDOMAbortError(),
-  });
-}
-
-// mimics TshdRpcError
-function newGrpcAbortError() {
-  return { code: 'CANCELLED' };
-}
